fix missed the HA case
