+++ tajo-catalog/tajo-catalog-client/src/main/java/org/apache/tajo/catalog/CatalogClient.java
@@ -24,28 +24,66 @@ import org.apache.tajo.conf.TajoConf;
 import org.apache.tajo.conf.TajoConf.ConfVars;
 import org.apache.tajo.rpc.NettyClientBase;
 import org.apache.tajo.rpc.RpcClientManager;
+import org.apache.tajo.rpc.RpcConstants;
+import org.apache.tajo.service.ServiceTracker;
+import org.apache.tajo.service.ServiceTrackerFactory;
 import org.apache.tajo.util.NetUtils;
 
 import java.io.IOException;
+import java.net.InetSocketAddress;
+import java.util.concurrent.TimeUnit;
 
 /**
  * CatalogClient provides a client API to access the catalog server.
  */
 public class CatalogClient extends AbstractCatalogClient {
+  protected NettyClientBase client;
+  protected ServiceTracker serviceTracker;
+  protected InetSocketAddress catalogServerAddr;
   /**
    * @throws java.io.IOException
    *
    */
   public CatalogClient(final TajoConf conf) throws IOException {
-    super(conf, NetUtils.createSocketAddr(conf.getVar(ConfVars.CATALOG_ADDRESS)));
+    super(conf);
+    this.catalogServerAddr = NetUtils.createSocketAddr(conf.getVar(ConfVars.CATALOG_ADDRESS));
+    this.serviceTracker = ServiceTrackerFactory.get(conf);
   }
 
-  public CatalogClient(TajoConf conf, String host, int port) throws IOException {
-    super(conf, NetUtils.createSocketAddr(host, port));
-  }
 
   @Override
   BlockingInterface getStub() throws ServiceException {
     return getCatalogConnection().getStub();
+  }
+
+  private InetSocketAddress getCatalogServerAddr() {
+    if (!conf.getBoolVar(TajoConf.ConfVars.TAJO_MASTER_HA_ENABLE)) {
+      return catalogServerAddr;
+    } else {
+      return serviceTracker.getCatalogAddress();
+    }
+  }
+
+  public synchronized NettyClientBase getCatalogConnection() throws ServiceException {
+    if (client != null && client.isConnected()) return client;
+    else {
+      try {
+        if (client != null && client.isConnected()) return client;
+        RpcClientManager.cleanup(client);
+
+        int retry = conf.getInt(RpcConstants.RPC_CLIENT_RETRY_MAX, RpcConstants.DEFAULT_RPC_RETRIES);
+        // Client do not closed on idle state for support high available
+        this.client = RpcClientManager.getInstance().newClient(getCatalogServerAddr(), CatalogProtocol.class, false,
+            retry, 0, TimeUnit.SECONDS, false);
+      } catch (Exception e) {
+        throw new ServiceException(e);
+      }
+      return client;
+    }
+  }
+
+  @Override
+  public void close() {
+    RpcClientManager.cleanup(client);
   }
 }
